Read the whole id or none, and split the wire at its two questions
The fallback matched a PREFIX, so an id from a source that is not this
one — a mirror row carries the incumbent's — could read "line 7 of the
export" as line 7, a number the reader would go looking for and not
find. That was already true of the parse this replaces; anchored, what
is left is an incumbent id spelled exactly `line 7` on a report stored
before the line was carried.

The file passed the 500-line cap, and the split it wanted was already
there: one half takes a caller's mapping apart and refuses what the
estate cannot receive, the other puts a finished run's outcome into a
shape a person reads. Nothing but history held them together.

Moving toContractImportReport made every line of it new to the strict
lint pass, which judges new code — and it was two over the complexity
ceiling. The two refused-row loops were the same loop written twice, so
they are one function that answers how many rows were new.
