+++ backend/internal/compose/csvimportwire.go
@@ -204,133 +204,6 @@ func toContractImportRun(run migration.Run) crmcontracts.ImportRun {
 	return out
 }
 
-func toContractImportReport(run migration.Run) crmcontracts.ImportRunReport {
-	out := crmcontracts.ImportRunReport{
-		RunId:  openapi_types.UUID(run.ID),
-		Status: crmcontracts.ImportRunStatus(run.Status),
-		Issues: []crmcontracts.ImportRowIssue{},
-	}
-	if run.Mapping != nil {
-		out.SourceKeyUsed = run.Mapping.SourceKey
-	}
-
-	// Predicted counts and actual ones are never summed: a finished run reports
-	// what it did, an awaiting one what it will do. Adding them would report
-	// twice the rows the file holds — and the stored report DOES carry both,
-	// because a run's own report is merged into the dry run's so a resumed
-	// attempt keeps what the earlier one already achieved.
-	committed := run.Status == migration.StatusComplete || run.Status == migration.StatusFailed ||
-		run.Status == migration.StatusUndoing || run.Status == migration.StatusUndone
-	seen := map[string]bool{}
-	duplicates := 0
-	for _, o := range run.Report.Objects {
-		out.RowsRead += o.MirrorCount
-		if committed {
-			out.Disposition.Created += o.Created
-			out.Disposition.Updated += o.Updated
-		} else {
-			out.Disposition.Created += o.WillCreate
-			out.Disposition.Updated += o.WillUpdate
-		}
-		// The commit's own count once there is one, for the same reason
-		// Created replaces WillCreate above: the estate moves between the
-		// preview and the approval — a colleague creates one of the companies,
-		// an earlier run lands it — and a finished report that stated the
-		// prediction would describe what was expected rather than what
-		// happened.
-		if committed {
-			duplicates += o.Duplicated
-		} else {
-			duplicates += o.WillDuplicate
-		}
-		for _, s := range o.Skipped {
-			// The same row skipped by the dry run and again by the commit is
-			// ONE row the human must go fix, named by its line.
-			// Keyed on the EXTERNAL ID, which is what identifies a row. Keyed
-			// on the line instead, every skipped row in a file that carries its
-			// own key column collapsed onto the first — the line was derived
-			// from the id's text shape back then and answered 0 for all of them
-			// — so two refused rows were reported as one skip and one phantom
-			// `unchanged`, and the four counts summed to less than rows_read.
-			// The line is carried now and would no longer collide, but the id
-			// is still the right key: it is what a row IS, where a line is
-			// where it happened to sit.
-			if seen[s.ExternalID] {
-				continue
-			}
-			seen[s.ExternalID] = true
-			out.Disposition.Skipped++
-			out.Issues = append(out.Issues, crmcontracts.ImportRowIssue{
-				Line:   lineOf(s),
-				Reason: s.Reason,
-			})
-		}
-		// A collision is reported the same way a skip is — by line, in the
-		// file's own terms — but it does NOT touch Disposition.Skipped: the
-		// row is counted in Created, and adding it here too would report two
-		// outcomes for one row and leave `unchanged` short by the difference.
-		for _, c := range o.Collisions {
-			if seen[c.ExternalID] {
-				continue
-			}
-			seen[c.ExternalID] = true
-			out.Issues = append(out.Issues, crmcontracts.ImportRowIssue{
-				Line:   lineOf(c),
-				Reason: c.Reason,
-			})
-		}
-	}
-
-	// Reported even when zero: "0 duplicates" is the answer to the question a
-	// person is asking before they approve, and an omitted field reads as "not
-	// checked" rather than "none found".
-	out.Disposition.Duplicates = &duplicates
-	out.Links = linksOf(run.Report, committed)
-
-	// A finished run's stored report can carry more outcomes than the file has
-	// rows, and there are two causes with different right answers.
-	//
-	// A row the DRY RUN skipped whose collision then vanished commits as a
-	// CREATE, and the stale skip is stored beside the real create — the engine
-	// folds attempts by object class and concatenates their skipped lists. Here
-	// the skip is the entry to drop.
-	//
-	// A RESUMED run can also double-count: ObjectReport.record runs before
-	// advanceCheckpoint, so a checkpoint that fails to persist leaves a counted
-	// row the resume walks again, and attempt reports add their counts. Here the
-	// surplus is in `created`/`updated`, and taking it out of `skipped` would
-	// erase a genuine refusal.
-	//
-	// Nothing in the stored report distinguishes them: per-row entries exist for
-	// skips alone, and everything else is an aggregate. Only the SECOND cause
-	// needs a resume to happen at all, so the attempt count is what tells them
-	// apart — a single-attempt run can only have the first.
-	if committed {
-		surplus := out.Disposition.Created + out.Disposition.Updated +
-			out.Disposition.Skipped - out.RowsRead
-		if surplus > 0 && len(run.Report.Objects) == 1 {
-			out.Disposition.Skipped = max(out.Disposition.Skipped-surplus, 0)
-		}
-	}
-
-	// Unchanged is DERIVED, never carried: it is the rows that were read and
-	// then neither created, updated nor skipped. Carrying the stored figure
-	// would add the dry run's count to the commit's and report more rows than
-	// the file holds — which is the contract's own invariant ("the four sum to
-	// the rows read") failing in the one place a human reads it.
-	out.Disposition.Unchanged = max(
-		out.RowsRead-out.Disposition.Created-out.Disposition.Updated-out.Disposition.Skipped,
-		0,
-	)
-	// Present once the run has been undone, not while undoing — a
-	// still-in-progress or interrupted reversal's partial counts are the
-	// run's own internal resume state, not a finished outcome to report.
-	if run.UndoReport != nil && run.Status == migration.StatusUndone {
-		out.Undo = toContractUndoReport(run.ID, run.Status, *run.UndoReport)
-	}
-	return out
-}
-
 // toContractUndoReport renders the reversal outcome (IEM-WIRE-9).
 func toContractUndoReport(id migration.RunID, status string, rep migration.UndoReport) *crmcontracts.ImportUndoReport {
 	kept := make([]struct {
@@ -362,68 +235,6 @@ func toContractUndoReport(id migration.RunID, status string, rep migration.UndoR
 		Kept:          kept,
 		Errored:       errored,
 	}
-}
-
-// linksOf renders the connections a run makes, for the half of the report that
-// is not about rows.
-//
-// The stored count means two different things either side of approval, which is
-// the engine's own shape rather than a quirk to smooth over. A dry run cannot
-// resolve an endpoint — it writes nothing and reads nothing about the other end
-// — so what it can honestly say is how many rows NAMED an employer. After the
-// commit the same field counts the edges actually written. Reporting them under
-// one number would let a preview promising 12,000 links be answered by 3,000
-// applied without either figure ever being wrong.
-//
-// Always non-nil for the same reason `duplicates` is: "0 links" answers the
-// question, an absent field reads as "not checked".
-func linksOf(rep *migration.Report, committed bool) *crmcontracts.ImportRunLinks {
-	links := crmcontracts.ImportRunLinks{}
-	if rep == nil {
-		empty := []crmcontracts.ImportUnresolvedLink{}
-		links.Unresolved = &empty
-		return &links
-	}
-	// One stored count, two meanings, decided by which side of approval the run
-	// is on. Before it, the engine has no writer and cannot resolve an endpoint,
-	// so the number is what the file asks for that CAN be linked. After it, the
-	// same field holds what was actually written. Reporting both under one name
-	// would let a preview promising 12,000 links be answered by 3,000 applied
-	// without either figure ever having been wrong.
-	if committed {
-		links.Applied = rep.Associations
-	}
-	// Offered is the whole ask either way: what landed, plus what named a company
-	// the run could not find.
-	links.Offered = rep.Associations + len(rep.AssociationsSkipped)
-	unresolved := make([]crmcontracts.ImportUnresolvedLink, 0, len(rep.AssociationsSkipped))
-	for _, s := range rep.AssociationsSkipped {
-		unresolved = append(unresolved, crmcontracts.ImportUnresolvedLink{
-			From: s.From, To: s.To, Reason: s.Reason,
-		})
-	}
-	links.Unresolved = &unresolved
-	return &links
-}
-
-// lineOf is the file line a skip names, from the row itself.
-//
-// The fallback is for reports STORED before the line was carried. Their JSON
-// has no `line` field, so it decodes as zero — and for the skips the source
-// disclosed, the id still spells "line N", which is where the line used to be
-// read from for every skip. Deriving it only when the carried one is absent
-// keeps those reports readable through a rollout without putting the old
-// parse back in the path: a row that carries its own id answers 0 there, which
-// is the defect this replaced.
-func lineOf(skip migration.SkippedRow) int {
-	if skip.Line != 0 {
-		return skip.Line
-	}
-	var line int
-	if _, err := fmt.Sscanf(skip.ExternalID, "line %d", &line); err != nil {
-		return 0
-	}
-	return line
 }
 
 // readImportUpload takes the multipart body apart under the deployment's import
